+++ Assets/Scripts/com.utkaka.InstancedVoxels/Runtime/Rendering/BrgRenderer/Metadata/PerInstanceMetadataValue.cs
@@ -4,14 +4,14 @@ namespace com.utkaka.InstancedVoxels.Runtime.Rendering.BrgRenderer.Metadata {
     public class PerInstanceMetadataValue<T> : BatchMetadataValue<T> where T : unmanaged {
         public PerInstanceMetadataValue(string shaderProperty) : base(shaderProperty) { }
 
-        public override long GetBufferSizeInFloat(uint instanceCount) {
+        public override int GetBufferSizeInFloat(int instanceCount) {
             return SizeInFloat * instanceCount;
         }
 
-        public override MetadataValue GetMetadataValue(ref uint offset, uint instanceCount) {
+        public override MetadataValue GetMetadataValue(ref int offset, int instanceCount) {
             var value = new MetadataValue {
                 NameID = Id,
-                Value = offset | 0x80000000
+                Value = (uint)offset | 0x80000000
             };
             offset += instanceCount * SizeInFloat;
             return value;
