+++ Assets/Scripts/com.utkaka.InstancedVoxels/Runtime/Rendering/BrgRenderer/Metadata/PerMaterialMetadataValue.cs
@@ -4,14 +4,14 @@ namespace com.utkaka.InstancedVoxels.Runtime.Rendering.BrgRenderer.Metadata {
     public class PerMaterialMetadataValue<T> : BatchMetadataValue<T> where T : unmanaged {
         public PerMaterialMetadataValue(string shaderProperty) : base(shaderProperty) { }
         
-        public override long GetBufferSizeInFloat(uint instanceCount) {
+        public override int GetBufferSizeInFloat(int instanceCount) {
             return SizeInFloat;
         }
         
-        public override MetadataValue GetMetadataValue(ref uint offset, uint instanceCount) {
+        public override MetadataValue GetMetadataValue(ref int offset, int instanceCount) {
             var value = new MetadataValue {
                 NameID = Id,
-                Value = offset
+                Value = (uint)offset
             };
             offset += SizeInFloat;
             return value;
